refactor : 도메인과 뷰에 알맞게 enum 재정의
