+++ src/main/java/blackjack/domain/CardNumber.java
@@ -1,26 +1,28 @@
 package blackjack.domain;
 
 public enum CardNumber {
-    ACE("A", 1),
-    TWO("2", 2),
-    THREE("3", 3),
-    FOUR("4", 4),
-    FIVE("5", 5),
-    SIX("6", 6),
-    SEVEN("7", 7),
-    EIGHT("8", 8),
-    NINE("9", 9),
-    TEN("10", 10),
-    JACK("J", 10),
-    QUEEN("Q", 10),
-    KING("K", 10),
+    ACE(1),
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8),
+    NINE(9),
+    TEN(10),
+    JACK(10),
+    QUEEN(10),
+    KING(10),
     ;
 
-    final String name;
     final int value;
 
-    CardNumber(String name, int value) {
-        this.name = name;
+    CardNumber(int value) {
         this.value = value;
+    }
+
+    public int getValue() {
+        return value;
     }
 }
